API Protection

+++ backend/app/api/routes.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from app.db.database import SessionLocal
 from app.models.analysis import Analysis
 from app.schemas.analysis import AnalysisCreate, AnalysisUpdate, AnalysisResponse
+from app.core.dependencies import get_current_user, get_admin_user
 from typing import List
 
 router = APIRouter(prefix="/api/v1/analyses", tags=["analyses"])
@@ -16,8 +17,8 @@ def get_db():
 
 # CREATE
 @router.post("/", response_model=AnalysisResponse)
-def create_analysis(data: AnalysisCreate, db: Session = Depends(get_db)):
-    new_analysis = Analysis(image_path=data.image_path, poles_number=data.poles_number, processing_time=data.processing_time)
+def create_analysis(data: AnalysisCreate, db: Session = Depends(get_db), user = Depends(get_current_user)):
+    new_analysis = Analysis(image_path=data.image_path, poles_number=data.poles_number, processing_time=data.processing_time, user_id=user.id)
     db.add(new_analysis)
     db.commit()
     db.refresh(new_analysis)
@@ -25,20 +26,31 @@ def create_analysis(data: AnalysisCreate, db: Session = Depends(get_db)):
 
 # READ ALL
 @router.get("/", response_model=List[AnalysisResponse])
-def get_analyses(db: Session = Depends(get_db)):
-    return db.query(Analysis).all()
+def get_analyses(db: Session = Depends(get_db), user = Depends(get_current_user)):
+    if user.role == "admin":
+        return db.query(Analysis).all()
+
+    return db.query(Analysis).filter(Analysis.user_id == user.id).all()
 
 # READ ONE
 @router.get("/{analysis_id}", response_model=AnalysisResponse)
-def get_analysis(analysis_id: int, db: Session = Depends(get_db)):
+def get_analysis(analysis_id: int, db: Session = Depends(get_db), user = Depends(get_current_user)):
     analysis = db.query(Analysis).filter(Analysis.id == analysis_id).first()
+
     if not analysis:
         raise HTTPException(status_code=404, detail="Analysis not found")
+
+    if user.role != "admin" and analysis.user_id != user.id:
+        raise HTTPException(status_code=403, detail="Access denied")
+
     return analysis
 
 # UPDATE
 @router.put("/{analysis_id}", response_model=AnalysisResponse)
-def update_analysis(analysis_id: int, data: AnalysisUpdate, db: Session = Depends(get_db)):
+def update_analysis(analysis_id: int, data: AnalysisUpdate, db: Session = Depends(get_db), user = Depends(get_current_user)):
+    if user.role != "admin":
+        raise HTTPException(status_code=403, detail="Only admin can update analysis")
+
     analysis = db.query(Analysis).filter(Analysis.id == analysis_id).first()
 
     if not analysis:
@@ -56,11 +68,14 @@ def update_analysis(analysis_id: int, data: AnalysisUpdate, db: Session = Depend
 
 # DELETE
 @router.delete("/{analysis_id}")
-def delete_analysis(analysis_id: int, db: Session = Depends(get_db)):
+def delete_analysis(analysis_id: int, db: Session = Depends(get_db), user = Depends(get_current_user)):
     analysis = db.query(Analysis).filter(Analysis.id == analysis_id).first()
 
     if not analysis:
         raise HTTPException(status_code=404, detail="Analysis not found")
+
+    if user.role != "admin" and analysis.user_id != user.id:
+        raise HTTPException(status_code=403, detail="Access denied")
 
     db.delete(analysis)
     db.commit()
